fix PR review changes
